+++ e2e/project/frontity.settings.ts
@@ -77,10 +77,10 @@ const settings: Settings = [
       "e2e-analytics",
       "@frontity/tiny-router",
       {
-        name: "@frontity/google-tag-manager",
+        name: "@frontity/google-tag-manager-analytics",
         state: {
-          googleTagManager: {
-            containerId: "UA-XXXXXX-X",
+          googleTagManagerAnalytics: {
+            containerId: "GTM-XXXXXX-X",
           },
         },
       },
@@ -120,6 +120,25 @@ const settings: Settings = [
       {
         name: "@frontity/wp-source",
         state: { source: { api: "https://domain.com/wp-json" } },
+      },
+    ],
+  },
+  {
+    name: "analytics",
+    packages: ["e2e-analytics", "@frontity/tiny-router", "@frontity/analytics"],
+  },
+  {
+    name: "google-analytics",
+    packages: [
+      "e2e-analytics",
+      "@frontity/tiny-router",
+      {
+        name: "@frontity/google-analytics",
+        state: {
+          googleAnalytics: {
+            trackingIds: ["UA-XXXXXX-X", "UA-YYYYYY-Y"],
+          },
+        },
       },
     ],
   },