+++ packages/tiny-router/CHANGELOG.md
@@ -12,6 +12,16 @@
   - @frontity/router@1.2.0-infinite-scroll-beta.0
   - frontity@1.8.1-infinite-scroll-beta.0
   - @frontity/source@1.2.2-infinite-scroll-beta.0
+## 1.2.1
+
+### Patch Changes
+
+- [`73f9ae35`](https://github.com/frontity/frontity/commit/73f9ae35a64e2dc8038fc1b21ee4f10147680530) [#494](https://github.com/frontity/frontity/pull/494) Thanks [@juanmaguitar](https://github.com/juanmaguitar)! - README created.
+
+- Updated dependencies [[`ba13f70a`](https://github.com/frontity/frontity/commit/ba13f70ae2a4360ca21c77aed1c920c02e9d45b8), [`62fce1e5`](https://github.com/frontity/frontity/commit/62fce1e5c117faeb5902dc0ddae3b13d95cd925b), [`3f61f711`](https://github.com/frontity/frontity/commit/3f61f71197d33b478427d1b74882c31258861e92), [`868c120f`](https://github.com/frontity/frontity/commit/868c120f2ede7a2f9013f6e659e1b0a1bf2785fe)]:
+  - frontity@1.9.0
+  - @frontity/source@1.2.2
+  - @frontity/router@1.1.1
 
 ## 1.2.0